Cleanup organise and refactor core module classes and api
