+++ src/core/storage/CwaVuexModule/actions/index.ts
@@ -1,0 +1,1 @@
+export { FetchComponentMetadata } from './FetchComponentMetadata'
