+++ src/core/storage/CwaVuexModule/mutations/index.ts
@@ -1,0 +1,6 @@
+export { TogglePublishable } from './TogglePublishable'
+export { DeleteResource } from './DeleteResource'
+export { SetResource } from './SetResource'
+export { SetCurrentRoute } from './SetCurrentRoute'
+export { MergeNewResources } from './MergeNewResources'
+export { ResetCurrentResources } from './ResetCurrentResources'
