+++ src/core/templates/components/admin/cwa-admin-bar-page-info-modal.vue
@@ -27,7 +27,7 @@ export default Vue.extend({
   data() {
     return {
       // @ts-ignore
-      iri: this.$cwa.currentPageIri
+      iri: this.$cwa.loadedPage
     }
   }
 })
