+++ src/core/templates/components/default/Form/FormView/Text.vue
@@ -31,7 +31,7 @@ import Wrapper from '@cwa/nuxt-module/core/templates/components/default/Form/For
 import {
   FormExtraSubmitData,
   FormView
-} from '@cwa/nuxt-module/core/vuex/FormsVuexModule'
+} from '@cwa/nuxt-module/core/storage/FormsVuexModule'
 
 export default Vue.extend({
   components: { Wrapper },
